+++ library/src/main/java/org/mustangproject/ZUGFeRD/ZUGFeRDImporter.java
@@ -262,7 +262,7 @@ public class ZUGFeRDImporter extends ZUGFeRDInvoiceImporter {
 	 * @return the Payment Terms
 	 */
 	public String getPaymentTerms() {
-		return extractString("//*[local-name() = 'SpecifiedTradePaymentTerms']//*[local-name() = 'Description']");
+		return importedInvoice.getPaymentTermDescription();
 	}
 
 	/**
@@ -659,6 +659,7 @@ public class ZUGFeRDImporter extends ZUGFeRDInvoiceImporter {
 	 *
 	 * @return a List of LineItem instances
 	 */
+	@Deprecated
 	public List<Item> getLineItemList() {
 		final List<Node> nodeList = getLineItemNodes();
 		final List<Item> lineItemList = new ArrayList<>();